requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.12.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
tiktoken==0.5.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
